// nextgen-simtool-ui/src/app/model/configParams.ts
export class ConfigParams {
  messageType: string;
  paths: [];
  searchKeywords: [];
  productName: string;
  isDefault: string;
  partnerId: string;
  contentType: string;
  payload: string;
  searchValue: string;

  // constructor(messageType: string,
  //             paths: [],
  //             searchKeywords: [],
  //             productName: string,
  //             isDefault: string,
  //             partnerId: string,
  //             contentType: string,
  //             payload: string)
  // {
  //   this.messageType = messageType;
  //   this.paths = paths;
  //   this.searchKeywords = searchKeywords;
  //   this.productName = productName;
  //   this.partnerId = partnerId;
  //   this.contentType = contentType;
  //   this.payload = payload;
  //   this.isDefault = isDefault;
  // }
}